% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFullScan.R
\name{runFullScan}
\alias{runFullScan}
\title{runFullScan}
\usage{
runFullScan(
  species,
  mods = NULL,
  UTRonly = TRUE,
  shadow = 15,
  cores = 8,
  maxLogKd = c(-0.3, -0.3),
  save.path = NULL,
  ...
)
}
\arguments{
\item{species}{}

\item{mods}{}

\item{UTRonly}{}

\item{shadow}{}

\item{cores}{}

\item{maxLogKd}{}

\item{save.path}{}

\item{...}{}
}
\description{
runFullScan
}
