% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanMiRAnno.R
\name{ScanMiRAnno-methods}
\alias{ScanMiRAnno-methods}
\alias{summary,ScanMiRAnno-method}
\alias{show,ScanMiRAnno-method}
\title{Methods for the \code{\link{ScanMiRAnno}} class}
\usage{
\S4method{summary}{ScanMiRAnno}(object)

\S4method{show}{ScanMiRAnno}(object)
}
\arguments{
\item{object}{An object of class \code{\link{ScanMiRAnno}}}
}
\value{
Depends on the method.
}
\description{
Methods for the \code{\link{ScanMiRAnno}} class
}
\seealso{
\code{\link{ScanMiRAnno}}
}
