% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanMiRserver.R
\name{scanMiRApp}
\alias{scanMiRApp}
\title{scanMiRApp
A wrapper for launching the scanMiRApp shiny app}
\usage{
scanMiRApp(annotations = NULL, ...)
}
\arguments{
\item{annotations}{A named list of \code{\link{ScanMiRAnno}} objects. If 
omitted, will use the base ones.}

\item{...}{Passed to \code{\link{scanMiRserver}}}
}
\value{
A shiny app
}
\description{
scanMiRApp
A wrapper for launching the scanMiRApp shiny app
}
