% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFullScan.R
\name{runFullScan}
\alias{runFullScan}
\title{runFullScan}
\usage{
runFullScan(
  annotation,
  mods = NULL,
  UTRonly = TRUE,
  shadow = 15,
  cores = 1,
  maxLogKd = c(-0.3, -0.3),
  save.path = NULL,
  ...
)
}
\arguments{
\item{annotation}{}

\item{mods}{}

\item{UTRonly}{}

\item{shadow}{}

\item{cores}{}

\item{maxLogKd}{}

\item{save.path}{}

\item{...}{}
}
\description{
runFullScan
}
\examples{
# not run
# anno <- ScanMiRAnno("Rnor_6")
# seq <- runFullScan( annotation=anno )
}
