% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFullScan.R
\name{runFullScan}
\alias{runFullScan}
\title{runFullScan}
\usage{
runFullScan(
  annotation,
  mods = NULL,
  UTRonly = TRUE,
  onlyCanonical = TRUE,
  shadow = 15,
  cores = 1,
  maxLogKd = c(-1, -1.5),
  save.path = NULL,
  ...
)
}
\arguments{
\item{annotation}{A \code{\link{ScanMiRAnno}} object}

\item{mods}{An optional `KdModelList` (defaults to the one in `annotation`)}

\item{UTRonly}{Whether to scan only UTRs}

\item{shadow}{The size of the ribosomal shadow at the UTR starts}

\item{cores}{The number of threads to use}

\item{maxLogKd}{The maximum log_kd of sites to report}

\item{save.path}{Optional, the path to which to save the results}

\item{...}{Arguments passed to `scanMiR::findSeedMatches`}
}
\description{
Runs a full miRNA scan on all protein-coding transcripts (or UTRs) of an
annotation.
}
\examples{
# not run
# anno <- ScanMiRAnno("Rnor_6")
# seq <- runFullScan( annotation=anno )
}
