% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanMiRserver.R
\name{scanMiRserver}
\alias{scanMiRserver}
\title{scanMiRserver}
\usage{
scanMiRserver(
  modlists,
  targetlists = list(),
  ensdbs = list(),
  genomes = list(),
  scans = list(),
  maxCacheSize = 100 * 10^6,
  BP = BiocParallel::SerialParam()
)
}
\arguments{
\item{modlists}{A named list of `KdModelList`}

\item{targetlists}{An optional list of aggregated targets, named as `modlists`}

\item{ensdbs}{A named list of `ensembldb` objects (with names corresponding 
to those of `modlists`)}

\item{genomes}{A named list of `BSgenome` objects (with names corresponding 
to those of `modlists`)}

\item{scans}{A named list of scan results (with names corresponding to 
those of `modlists`)}

\item{maxCacheSize}{Maximum cache size in bytes.}
}
\value{
A shiny server function
}
\description{
Server function for the scanMiR shiny app.
}
