% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanMiRserver.R
\name{scanMiRserver}
\alias{scanMiRserver}
\title{scanMiRserver}
\usage{
scanMiRserver(
  annotations = list(),
  modlists = NULL,
  maxCacheSize = 100 * 10^6,
  BP = SerialParam()
)
}
\arguments{
\item{annotations}{A named list of \code{\link{ScanMiRAnno}} object.}

\item{modlists}{A named list of `KdModelList` objects. If omitted, will 
fetch it from the annotation objects.}

\item{maxCacheSize}{Maximum cache size in bytes.}

\item{BP}{BPPARAM for multithreading}
}
\value{
A shiny server function
}
\description{
Server function for the scanMiR shiny app.
}
